fix: upload export warn
