+++ src/upload/interface.ts
@@ -1,4 +1,3 @@
-import { PropType } from 'vue';
 import { TdUploadProps, UploadRemoveContext } from './type';
 import { GlobalConfigProvider } from '../config-provider/type';
 import { TNode } from '../common';
@@ -27,25 +26,3 @@ export interface CommonDisplayFileProps {
 }
 
 export type UploadProps = TdUploadProps;
-
-export const commonProps = {
-  files: Array as PropType<CommonDisplayFileProps['files']>,
-  toUploadFiles: Array as PropType<CommonDisplayFileProps['toUploadFiles']>,
-  displayFiles: Array as PropType<CommonDisplayFileProps['displayFiles']>,
-  theme: String as PropType<CommonDisplayFileProps['theme']>,
-  abridgeName: Array as PropType<CommonDisplayFileProps['abridgeName']>,
-  placeholder: String,
-  classPrefix: String,
-  tips: String,
-  status: String as PropType<CommonDisplayFileProps['status']>,
-  locale: Object as PropType<CommonDisplayFileProps['locale']>,
-  sizeOverLimitMessage: String,
-  autoUpload: Boolean,
-  disabled: Boolean,
-  uploading: Boolean,
-  tipsClasses: [String, Array, Object] as PropType<CommonDisplayFileProps['tipsClasses']>,
-  errorClasses: [String, Array, Object] as PropType<CommonDisplayFileProps['errorClasses']>,
-  default: Function as PropType<CommonDisplayFileProps['default']>,
-  fileListDisplay: Function as PropType<CommonDisplayFileProps['fileListDisplay']>,
-  onRemove: Function as PropType<CommonDisplayFileProps['onRemove']>,
-};
